# server/openapi.yaml
openapi: 3.1.0
info:
  title: BharatGig API
  version: 0.1.0
  description: BharatGig Freelance Marketplace API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: getJobs
      tags: [jobs]
      summary: List job postings
      responses:
        "200":
          description: List of jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Post a new job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobRequest"
      responses:
        "201":
          description: Created job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /payments/create-order:
    post:
      operationId: createPaymentOrder
      tags: [payments]
      summary: Create a Razorpay order for escrow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount: { type: number }
                clientId: { type: string }
                freelancerId: { type: string }
                jobId: { type: string }
      responses:
        "201":
          description: Order created
...
  /payments/release/{escrowId}:
    post:
      operationId: releaseFunds
      tags: [payments]
      summary: Release escrowed funds to freelancer
      parameters:
        - name: escrowId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Funds released
components:
  schemas:
...
    Job:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: string }
        budget: { type: number }
        clientId: { type: string }
    CreateJobRequest:
      type: object
      required: [title, description, budget]
      properties:
        title: { type: string }
        description: { type: string }
        budget: { type: number }